educonnect/backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-decouple==3.8
django-cors-headers==4.7.0
celery==5.5.3
redis==6.4.0
beautifulsoup4==4.13.5
scrapy==2.13.3
selenium==4.35.0
requests==2.32.5
openai==1.102.0
python-dotenv==1.1.1
Pillow==11.3.0
djangorestframework-simplejwt==5.5.1
django-filter==25.1